+++ .gitignore
@@ -4,3 +4,4 @@ swear_counts.json
 quote_bank.json
 __pycache__/
 *.pyc
+activity_valut.json
