+++ modules/quote_keeper.py
